refactor: move home-page.ts into lib/content directory
