+++ pages/index.tsx
@@ -9,8 +9,8 @@ import { HomePageWhatsAppCTA } from "~/components/home/homepage-whatsapp-cta";
 import { Page } from "~/components/layout/page";
 import { Alert } from "~/components/ui/alert";
 import { Container } from "~/components/ui/container";
+import { attributes, html } from "~/lib/content/home-page";
 import siteConfig from "~/lib/content/site-config";
-import { attributes, html } from "~/lib/home-page";
 import { htmrTransform } from "~/lib/htmr-transformers";
 
 import { ClockIcon } from "@heroicons/react/outline";
